requirements.txt
pyfiglet
streamlits